v 1.0.2 move healthcheck to main, del extra code
