+++ app/_resources.py
@@ -1,4 +1,4 @@
 __build__ = '1.0'
-__version__ = '1.0.1'
+__version__ = '1.0.2'
 __author__ = 'Deskent'
 __appname__ = "Nirax Transceiver"
