+++ docker-swarm.dev.yml
@@ -14,7 +14,7 @@ services:
     volumes:
       - ./logs:/app/logs
     healthcheck:
-      test: curl --fail http://127.0.0.1:8000/api/transceiver/healthcheck || exit 1
+      test: curl --fail http://127.0.0.1:8000/healthcheck || exit 1
       interval: 30s
       retries: 5
       start_period: 20s
